# Task2/arkpz-pzpi-22-3-shpak-denys-task2/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Specification",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/analytics/{company_id}/optimal-back-route": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the optimal back route for the given route ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get optimal back route",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company_id",
                        "name": "company_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/analytics/{company_id}/optimal-route": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the optimal route for the given route ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get optimal route",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company_id",
                        "name": "company_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/analytics/{company_id}/weather-alert": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the weather alert for the given route ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get weather alert",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company_id",
                        "name": "company_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/routes": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new route with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "route"
                ],
                "summary": "Create a new route",
                "parameters": [
                    {
                        "description": "Route details",
                        "name": "route",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateRouteRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/routes/{route_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a route with the given ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "route"
                ],
                "summary": "Get a route",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Route ID",
                        "name": "route_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing route with the given ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "route"
                ],
                "summary": "Update an existing route",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Route ID",
                        "name": "route_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated route details",
                        "name": "route",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateRouteRequest"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a route with the given ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "route"
                ],
                "summary": "Delete a route",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Route ID",
                        "name": "route_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateRouteRequest": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "example": "Route 1"
                },
                "status": {
                    "type": "string",
                    "example": "not_started | active | completed"
                }
            }
        },
...
        "handlers.UpdateRouteRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Route 1"
                },
                "status": {
                    "type": "string",
                    "example": "not_started | active | completed"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Provide the Bearer token in the format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}